changed how we process purchase date ranges; first we ask a user for a starting date and and ending date to generate purchases between; in generatePurchases, we iterate day by day and generate a random amount of purchases for the current day (amount of purchases range between 60 and 100); when we reach the alloted purchase amount for the day, we iterate foward by 1 day and generate another random amount of purchases until we reach the end date; I think this method is better than the old method because (1) purchase days are now ordered by ID (which I think makes more sense, why would purchaseId 1 have a date 5/1/2005, but then purchaseId 2 have a date 4/20/2005, etc.) and (2) now we can specify ranges of dates and potentially be able to better control for trends when we actually start making tuples for the database

+++ src/com/store/data/generator/models/Purchase.java
@@ -10,7 +10,7 @@ public class Purchase
         return id;
     }
 
-    public int getSellingPrice()
+    public double getSellingPrice()
     {
         return sellingPrice;
     }
@@ -45,7 +45,7 @@ public class Purchase
     }
 
     public Purchase(final long id,
-                    final int sellingPrice,
+                    final double sellingPrice,
                     final DateTime date,
                     final long itemId,
                     final long employeeId)
@@ -58,7 +58,7 @@ public class Purchase
     }
 
     private final long id;
-    private final int sellingPrice;
+    private final double sellingPrice;
     private final DateTime date;
     private final long itemId;
     private final long employeeId;
